pyproject.toml
[tool.poetry]
name = "psychology"
version = "0.1.0"
description = ""
authors = ["Argenis Vargas <argenis1818@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
safety = "^2.3.5"
pydantic = "^2.5.1"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
pytest = "^7.4.3"
pytest-django = "^4.7.0"
pytest-mock = "^3.12.0"
pillow = "^10.1.0"
dj-database-url = "^2.1.0"
psycopg2 = "^2.9.9"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
/(
  | migrations
)/poetry config http-basic.

'''

[tool.coverage.run]
omit = [
    "apps/*/migrations/*",
]
